+++ appEditors/AppExcEditor.py
@@ -2028,7 +2028,22 @@ class AppExcEditor(QtCore.QObject):
 
         self.slot_array_frame.hide()
 
-        self.tools_box.addStretch()
+        layout.addStretch()
+
+        # Editor
+        self.exit_editor_button = QtWidgets.QPushButton(_('Exit Editor'))
+        self.exit_editor_button.setToolTip(
+            _("Exit from Editor.")
+        )
+        self.exit_editor_button.setStyleSheet("""
+                                      QPushButton
+                                      {
+                                          font-weight: bold;
+                                      }
+                                      """)
+        layout.addWidget(self.exit_editor_button)
+
+        self.exit_editor_button.clicked.connect(lambda: self.app.editor2object())
 
         # ## Toolbar events and properties
         self.tools_exc = {